requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-renderer-xlsx==0.4.1
et-xmlfile==1.1.0
gunicorn==20.0.4
openpyxl==3.0.7
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
